+++ src/extension.ts
@@ -3,7 +3,7 @@ import * as vscode from "vscode"
 import { markdownLinkCompletionProvider } from "./autocomplete-links/markdown-link-provider"
 import { fileDeletedHandler } from "./file-deleted/file-deleted-handler"
 import { fileRenamedHandler } from "./file-renamed/file-renamed-handler"
-import { removeLeadingPounds } from "./helpers/remove-leading-pounds"
+import { renameTitle } from "./rename-title/rename-title"
 
 export function activate(context: vscode.ExtensionContext): void {
   // autocomplete links by typing `[`
@@ -22,19 +22,4 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // rename document title
   context.subscriptions.push(vscode.commands.registerCommand("markdownIDE.renameDocumentTitle", renameTitle))
-}
-
-export async function renameTitle(): Promise<void> {
-  const titleLine = vscode.window.activeTextEditor?.document.lineAt(0)
-  if (!titleLine) {
-    // document doesn't have content
-    return
-  }
-  const oldTitle = removeLeadingPounds(titleLine.text)
-  const newTitle = await vscode.window.showInputBox({ value: oldTitle })
-  if (newTitle === undefined) {
-    // user aborted with ESC
-    return
-  }
-  console.log(newTitle)
 }
